+++ recipes/deploy-all-in-one.rb
@@ -167,6 +167,9 @@ deploy_revision "opencast" do
     install_aws_s3_distribution_service_config(most_recent_deploy, enable_s3, region, s3_distribution_bucket_name, s3_distribution_base_url)
     install_search_content_service_config(most_recent_deploy, search_content_enabled, region, s3_distribution_bucket_name, stack_name, search_content_index_url, search_content_lambda_name)
 #    install_opencast_images_properties(most_recent_deploy)
+    # OPC-139 Oauth config (for Engage)
+    install_oauthconsumerdetails_service_config(most_recent_deploy)
+
     # /all-in-one SPECIFIC
 
     if using_local_distribution